requirements.txt
streamlit
requests
pandas 
pymongo
psycopg2-binary
sqlalchemy

